requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.18
langchain-ollama==1.0.1
langsmith==0.7.16
ollama==0.6.1
orjson==3.11.7
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
